+++ lib/eve_online.rb
@@ -12,6 +12,7 @@ require 'eve_online/exceptions/forbidden'
 require 'eve_online/exceptions/internal_server_error'
 require 'eve_online/exceptions/bad_gateway'
 require 'eve_online/exceptions/service_unavailable'
+require 'eve_online/exceptions/no_content'
 
 # ESI API
 require 'eve_online/esi/base'