+++ lib/eve_online/exceptions/no_content.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module EveOnline
+  module Exceptions
+    class NoContent < Base
+    end
+  end
+end